sllIN: fix compiler warning

+++ examples/device/sllin/src/usb_descriptors.c
@@ -180,7 +180,7 @@ uint16_t const* tud_descriptor_string_cb(uint8_t index, uint16_t langid)
 		usb_get_desc_string(index, string_buffer, &capa_len);
 
 		// convert to UTF-16
-		for (int i = 0; i < capa_len; ++i) {
+		for (size_t i = 0; i < capa_len; ++i) {
 			wstring_buffer[i+1] = string_buffer[i];
 		}
